KAFKA-14588 [3/N] ConfigCommandTest rewritten in java
