+++ core/src/test/scala/unit/kafka/admin/ConfigCommandTest.scala
@@ -20,7 +20,7 @@ import java.util
 import java.util.Properties
 import kafka.admin.ConfigCommand.ConfigCommandOptions
 import kafka.cluster.Broker
-import kafka.utils.{Logging, TestUtils}
+import kafka.utils.Logging
 import kafka.zk.{AdminZkClient, KafkaZkClient}
 import org.apache.kafka.clients.admin._
 import org.apache.kafka.common.Node
@@ -30,11 +30,10 @@ import org.apache.kafka.common.internals.KafkaFutureImpl
 import org.apache.kafka.common.quota.{ClientQuotaAlteration, ClientQuotaFilter}
 import org.apache.kafka.common.security.scram.internals.ScramCredentialUtils
 import org.apache.kafka.common.utils.Sanitizer
-import org.apache.kafka.server.config.{ConfigType, ZooKeeperInternals}
+import org.apache.kafka.server.config.ZooKeeperInternals
 import org.junit.jupiter.api.Assertions._
 import org.junit.jupiter.api.Test
-import org.mockito.ArgumentMatchers.anyString
-import org.mockito.Mockito.{mock, times, verify, when}
+import org.mockito.Mockito.{mock, verify, when}
 
 import scala.collection.{Seq, mutable}
 import scala.jdk.CollectionConverters._
@@ -43,455 +42,6 @@ class ConfigCommandTest extends Logging {
 
   private val zkConnect = "localhost:2181"
   private val dummyAdminZkClient = new DummyAdminZkClient(null)
-
-  @Test
-  def shouldAddTopicConfigUsingZookeeper(): Unit = {
-    val createOpts = new ConfigCommandOptions(Array("--zookeeper", zkConnect,
-      "--entity-name", "my-topic",
-      "--entity-type", "topics",
-      "--alter",
-      "--add-config", "a=b,c=d"))
-
-    val zkClient = mock(classOf[KafkaZkClient])
-    when(zkClient.getEntityConfigs(anyString(), anyString())).thenReturn(new Properties())
-
-    class TestAdminZkClient(zkClient: KafkaZkClient) extends AdminZkClient(zkClient) {
-      override def changeTopicConfig(topic: String, configChange: Properties): Unit = {
-        assertEquals("my-topic", topic)
-        assertEquals("b", configChange.get("a"))
-        assertEquals("d", configChange.get("c"))
-      }
-    }
-
-    ConfigCommand.alterConfigWithZk(null, createOpts, new TestAdminZkClient(zkClient))
-  }
-
-  @Test
-  def shouldAlterTopicConfig(): Unit = {
-    doShouldAlterTopicConfig(false)
-  }
-
-  @Test
-  def shouldAlterTopicConfigFile(): Unit = {
-    doShouldAlterTopicConfig(true)
-  }
-
-  def doShouldAlterTopicConfig(file: Boolean): Unit = {
-    var filePath = ""
-    val addedConfigs = Map("delete.retention.ms" -> "1000000", "min.insync.replicas" -> "2")
-    if (file) {
-      val file = TestUtils.tempPropertiesFile(addedConfigs)
-      filePath = file.getPath
-    }
-
-    val resourceName = "my-topic"
-    val alterOpts = new ConfigCommandOptions(Array("--bootstrap-server", "localhost:9092",
-      "--entity-name", resourceName,
-      "--entity-type", "topics",
-      "--alter",
-      if (file) "--add-config-file" else "--add-config",
-      if (file) filePath else addedConfigs.map{case (k, v) => k + "=" + v}.mkString(","),
-      "--delete-config", "unclean.leader.election.enable"))
-    var alteredConfigs = false
-
-    def newConfigEntry(name: String, value: String): ConfigEntry =
-      ConfigTest.newConfigEntry(name, value, ConfigEntry.ConfigSource.DYNAMIC_TOPIC_CONFIG, false, false, List.empty[ConfigEntry.ConfigSynonym].asJava)
-
-    val resource = new ConfigResource(ConfigResource.Type.TOPIC, resourceName)
-    val configEntries = List(newConfigEntry("min.insync.replicas", "1"), newConfigEntry("unclean.leader.election.enable", "1")).asJava
-    val future = new KafkaFutureImpl[util.Map[ConfigResource, Config]]
-    future.complete(util.Collections.singletonMap(resource, new Config(configEntries)))
-    val describeResult: DescribeConfigsResult = mock(classOf[DescribeConfigsResult])
-    when(describeResult.all()).thenReturn(future)
-
-    val alterFuture = new KafkaFutureImpl[Void]
-    alterFuture.complete(null)
-    val alterResult: AlterConfigsResult = mock(classOf[AlterConfigsResult])
-    when(alterResult.all()).thenReturn(alterFuture)
-
-    val node = new Node(1, "localhost", 9092)
-    val mockAdminClient = new MockAdminClient(util.Collections.singletonList(node), node) {
-      override def describeConfigs(resources: util.Collection[ConfigResource], options: DescribeConfigsOptions): DescribeConfigsResult = {
-        assertFalse(options.includeSynonyms(), "Config synonyms requested unnecessarily")
-        assertEquals(1, resources.size)
-        val resource = resources.iterator.next
-        assertEquals(resource.`type`, ConfigResource.Type.TOPIC)
-        assertEquals(resource.name, resourceName)
-        describeResult
-      }
-
-      override def incrementalAlterConfigs(configs: util.Map[ConfigResource, util.Collection[AlterConfigOp]], options: AlterConfigsOptions): AlterConfigsResult = {
-        assertEquals(1, configs.size)
-        val entry = configs.entrySet.iterator.next
-        val resource = entry.getKey
-        val alterConfigOps = entry.getValue
-        assertEquals(ConfigResource.Type.TOPIC, resource.`type`)
-        assertEquals(3, alterConfigOps.size)
-
-        val expectedConfigOps = Set(
-          new AlterConfigOp(newConfigEntry("delete.retention.ms", "1000000"), AlterConfigOp.OpType.SET),
-          new AlterConfigOp(newConfigEntry("min.insync.replicas", "2"), AlterConfigOp.OpType.SET),
-          new AlterConfigOp(newConfigEntry("unclean.leader.election.enable", ""), AlterConfigOp.OpType.DELETE)
-        )
-        assertEquals(expectedConfigOps.size, alterConfigOps.size)
-        expectedConfigOps.foreach { expectedOp =>
-          val actual = alterConfigOps.asScala.find(_.configEntry.name == expectedOp.configEntry.name)
-          assertNotEquals(actual, None)
-          assertEquals(expectedOp.opType, actual.get.opType)
-          assertEquals(expectedOp.configEntry.name, actual.get.configEntry.name)
-          assertEquals(expectedOp.configEntry.value, actual.get.configEntry.value)
-        }
-        alteredConfigs = true
-        alterResult
-      }
-    }
-    ConfigCommand.alterConfig(mockAdminClient, alterOpts)
-    assertTrue(alteredConfigs)
-    verify(describeResult).all()
-  }
-
-  @Test
-  def shouldDescribeConfigSynonyms(): Unit = {
-    val resourceName = "my-topic"
-    val describeOpts = new ConfigCommandOptions(Array("--bootstrap-server", "localhost:9092",
-      "--entity-name", resourceName,
-      "--entity-type", "topics",
-      "--describe",
-      "--all"))
-
-    val resource = new ConfigResource(ConfigResource.Type.TOPIC, resourceName)
-    val future = new KafkaFutureImpl[util.Map[ConfigResource, Config]]
-    future.complete(util.Collections.singletonMap(resource, new Config(util.Collections.emptyList[ConfigEntry])))
-    val describeResult: DescribeConfigsResult = mock(classOf[DescribeConfigsResult])
-    when(describeResult.all()).thenReturn(future)
-
-    val node = new Node(1, "localhost", 9092)
-    val mockAdminClient = new MockAdminClient(util.Collections.singletonList(node), node) {
-      override def describeConfigs(resources: util.Collection[ConfigResource], options: DescribeConfigsOptions): DescribeConfigsResult = {
-        assertTrue(options.includeSynonyms(), "Synonyms not requested")
-        assertEquals(Set(resource), resources.asScala.toSet)
-        describeResult
-      }
-    }
-    ConfigCommand.describeConfig(mockAdminClient, describeOpts)
-    verify(describeResult).all()
-  }
-
-  @Test
-  def shouldNotAllowAddBrokerQuotaConfigWhileBrokerUpUsingZookeeper(): Unit = {
-    val alterOpts = new ConfigCommandOptions(Array("--zookeeper", zkConnect,
-      "--entity-name", "1",
-      "--entity-type", "brokers",
-      "--alter",
-      "--add-config", "leader.replication.throttled.rate=10,follower.replication.throttled.rate=20"))
-
-    val mockZkClient: KafkaZkClient = mock(classOf[KafkaZkClient])
-    val mockBroker: Broker = mock(classOf[Broker])
-    when(mockZkClient.getBroker(1)).thenReturn(Option(mockBroker))
-
-    assertThrows(classOf[IllegalArgumentException],
-      () => ConfigCommand.alterConfigWithZk(mockZkClient, alterOpts, dummyAdminZkClient))
-  }
-
-  @Test
-  def shouldNotAllowDescribeBrokerWhileBrokerUpUsingZookeeper(): Unit = {
-    val describeOpts = new ConfigCommandOptions(Array("--zookeeper", zkConnect,
-      "--entity-name", "1",
-      "--entity-type", "brokers",
-      "--describe"))
-
-    val mockZkClient: KafkaZkClient = mock(classOf[KafkaZkClient])
-    val mockBroker: Broker = mock(classOf[Broker])
-    when(mockZkClient.getBroker(1)).thenReturn(Option(mockBroker))
-
-    assertThrows(classOf[IllegalArgumentException],
-      () => ConfigCommand.describeConfigWithZk(mockZkClient, describeOpts, dummyAdminZkClient))
-  }
-
-  @Test
-  def shouldSupportDescribeBrokerBeforeBrokerUpUsingZookeeper(): Unit = {
-    val describeOpts = new ConfigCommandOptions(Array("--zookeeper", zkConnect,
-      "--entity-name", "1",
-      "--entity-type", "brokers",
-      "--describe"))
-
-    class TestAdminZkClient(zkClient: KafkaZkClient) extends AdminZkClient(zkClient) {
-      override def fetchEntityConfig(rootEntityType: String, sanitizedEntityName: String): Properties = {
-        assertEquals("brokers", rootEntityType)
-        assertEquals("1", sanitizedEntityName)
-
-        new Properties()
-      }
-    }
-
-    val mockZkClient: KafkaZkClient = mock(classOf[KafkaZkClient])
-    when(mockZkClient.getBroker(1)).thenReturn(None)
-
-    ConfigCommand.describeConfigWithZk(mockZkClient, describeOpts, new TestAdminZkClient(null))
-  }
-
-  @Test
-  def shouldAddBrokerLoggerConfig(): Unit = {
-    val node = new Node(1, "localhost", 9092)
-    verifyAlterBrokerLoggerConfig(node, "1", "1", List(
-      new ConfigEntry("kafka.log.LogCleaner", "INFO"),
-      new ConfigEntry("kafka.server.ReplicaManager", "INFO"),
-      new ConfigEntry("kafka.server.KafkaApi", "INFO")
-    ))
-  }
-
-  @Test
-  def testNoSpecifiedEntityOptionWithDescribeBrokersInZKIsAllowed(): Unit = {
-    val optsList = List("--zookeeper", zkConnect,
-      "--entity-type", ConfigType.BROKER,
-      "--describe"
-    )
-
-    new ConfigCommandOptions(optsList.toArray).checkArgs()
-  }
-
-  @Test
-  def testNoSpecifiedEntityOptionWithDescribeBrokersInBootstrapServerIsAllowed(): Unit = {
-    val optsList = List("--bootstrap-server", "localhost:9092",
-      "--entity-type", ConfigType.BROKER,
-      "--describe"
-    )
-
-    new ConfigCommandOptions(optsList.toArray).checkArgs()
-  }
-
-  @Test
-  def testDescribeAllBrokerConfig(): Unit = {
-    val optsList = List("--bootstrap-server", "localhost:9092",
-      "--entity-type", ConfigType.BROKER,
-      "--entity-name", "1",
-      "--describe",
-      "--all")
-
-    new ConfigCommandOptions(optsList.toArray).checkArgs()
-  }
-
-  @Test
-  def testDescribeAllTopicConfig(): Unit = {
-    val optsList = List("--bootstrap-server", "localhost:9092",
-      "--entity-type", ConfigType.TOPIC,
-      "--entity-name", "foo",
-      "--describe",
-      "--all")
-
-    new ConfigCommandOptions(optsList.toArray).checkArgs()
-  }
-
-  @Test
-  def testDescribeAllBrokerConfigBootstrapServerRequired(): Unit = {
-    val optsList = List("--zookeeper", zkConnect,
-      "--entity-type", ConfigType.BROKER,
-      "--entity-name", "1",
-      "--describe",
-      "--all")
-
-    assertThrows(classOf[IllegalArgumentException], () => new ConfigCommandOptions(optsList.toArray).checkArgs())
-  }
-
-  @Test
-  def testEntityDefaultOptionWithDescribeBrokerLoggerIsNotAllowed(): Unit = {
-    val optsList = List("--bootstrap-server", "localhost:9092",
-      "--entity-type", ConfigCommand.BrokerLoggerConfigType,
-      "--entity-default",
-      "--describe"
-    )
-
-    assertThrows(classOf[IllegalArgumentException], () => new ConfigCommandOptions(optsList.toArray).checkArgs())
-  }
-
-  @Test
-  def testEntityDefaultOptionWithAlterBrokerLoggerIsNotAllowed(): Unit = {
-    val optsList = List("--bootstrap-server", "localhost:9092",
-      "--entity-type", ConfigCommand.BrokerLoggerConfigType,
-      "--entity-default",
-      "--alter",
-      "--add-config", "kafka.log.LogCleaner=DEBUG"
-    )
-
-    assertThrows(classOf[IllegalArgumentException], () => new ConfigCommandOptions(optsList.toArray).checkArgs())
-  }
-
-  @Test
-  def shouldRaiseInvalidConfigurationExceptionWhenAddingInvalidBrokerLoggerConfig(): Unit = {
-    val node = new Node(1, "localhost", 9092)
-    // verifyAlterBrokerLoggerConfig tries to alter kafka.log.LogCleaner, kafka.server.ReplicaManager and kafka.server.KafkaApi
-    // yet, we make it so DescribeConfigs returns only one logger, implying that kafka.server.ReplicaManager and kafka.log.LogCleaner are invalid
-    assertThrows(classOf[InvalidConfigurationException], () => verifyAlterBrokerLoggerConfig(node, "1", "1", List(
-      new ConfigEntry("kafka.server.KafkaApi", "INFO")
-    )))
-  }
-
-  @Test
-  def shouldAddDefaultBrokerDynamicConfig(): Unit = {
-    val node = new Node(1, "localhost", 9092)
-    verifyAlterBrokerConfig(node, "", List("--entity-default"))
-  }
-
-  @Test
-  def shouldAddBrokerDynamicConfig(): Unit = {
-    val node = new Node(1, "localhost", 9092)
-    verifyAlterBrokerConfig(node, "1", List("--entity-name", "1"))
-  }
-
-  def verifyAlterBrokerConfig(node: Node, resourceName: String, resourceOpts: List[String]): Unit = {
-    val optsList = List("--bootstrap-server", "localhost:9092",
-      "--entity-type", "brokers",
-      "--alter",
-      "--add-config", "message.max.bytes=10,leader.replication.throttled.rate=10") ++ resourceOpts
-    val alterOpts = new ConfigCommandOptions(optsList.toArray)
-    val brokerConfigs = mutable.Map[String, String]("num.io.threads" -> "5")
-
-    val resource = new ConfigResource(ConfigResource.Type.BROKER, resourceName)
-    val configEntries = util.Collections.singletonList(new ConfigEntry("num.io.threads", "5"))
-    val future = new KafkaFutureImpl[util.Map[ConfigResource, Config]]
-    future.complete(util.Collections.singletonMap(resource, new Config(configEntries)))
-    val describeResult: DescribeConfigsResult = mock(classOf[DescribeConfigsResult])
-    when(describeResult.all()).thenReturn(future)
-
-    val alterFuture = new KafkaFutureImpl[Void]
-    alterFuture.complete(null)
-    val alterResult: AlterConfigsResult = mock(classOf[AlterConfigsResult])
-    when(alterResult.all()).thenReturn(alterFuture)
-
-    val mockAdminClient = new MockAdminClient(util.Collections.singletonList(node), node) {
-      override def describeConfigs(resources: util.Collection[ConfigResource], options: DescribeConfigsOptions): DescribeConfigsResult = {
-        assertFalse(options.includeSynonyms(), "Config synonyms requested unnecessarily")
-        assertEquals(1, resources.size)
-        val resource = resources.iterator.next
-        assertEquals(ConfigResource.Type.BROKER, resource.`type`)
-        assertEquals(resourceName, resource.name)
-        describeResult
-      }
-
-      override def alterConfigs(configs: util.Map[ConfigResource, Config], options: AlterConfigsOptions): AlterConfigsResult = {
-        assertEquals(1, configs.size)
-        val entry = configs.entrySet.iterator.next
-        val resource = entry.getKey
-        val config = entry.getValue
-        assertEquals(ConfigResource.Type.BROKER, resource.`type`)
-        config.entries.forEach { e => brokerConfigs.put(e.name, e.value) }
-        alterResult
-      }
-    }
-    ConfigCommand.alterConfig(mockAdminClient, alterOpts)
-    assertEquals(Map("message.max.bytes" -> "10", "num.io.threads" -> "5", "leader.replication.throttled.rate" -> "10"),
-      brokerConfigs.toMap)
-    verify(describeResult).all()
-  }
-
-  @Test
-  def shouldDescribeConfigBrokerWithoutEntityName(): Unit = {
-    val describeOpts = new ConfigCommandOptions(Array("--bootstrap-server", "localhost:9092",
-      "--entity-type", "brokers",
-      "--describe"))
-
-    val BrokerDefaultEntityName = ""
-    val resourceCustom = new ConfigResource(ConfigResource.Type.BROKER, "1")
-    val resourceDefault = new ConfigResource(ConfigResource.Type.BROKER, BrokerDefaultEntityName)
-    val future = new KafkaFutureImpl[util.Map[ConfigResource, Config]]
-    val emptyConfig = new Config(util.Collections.emptyList[ConfigEntry])
-    val resultMap = Map(resourceCustom -> emptyConfig, resourceDefault -> emptyConfig).asJava
-    future.complete(resultMap)
-    val describeResult: DescribeConfigsResult = mock(classOf[DescribeConfigsResult])
-    // make sure it will be called 2 times: (1) for broker "1" (2) for default broker ""
-    when(describeResult.all()).thenReturn(future)
-
-    val node = new Node(1, "localhost", 9092)
-    val mockAdminClient = new MockAdminClient(util.Collections.singletonList(node), node) {
-      override def describeConfigs(resources: util.Collection[ConfigResource], options: DescribeConfigsOptions): DescribeConfigsResult = {
-        assertTrue(options.includeSynonyms(), "Synonyms not requested")
-        val resource = resources.iterator.next
-        assertEquals(ConfigResource.Type.BROKER, resource.`type`)
-        assertTrue(resourceCustom.name == resource.name || resourceDefault.name == resource.name)
-        assertEquals(1, resources.size)
-        describeResult
-      }
-    }
-    ConfigCommand.describeConfig(mockAdminClient, describeOpts)
-    verify(describeResult, times(2)).all()
-  }
-
-  private def verifyAlterBrokerLoggerConfig(node: Node, resourceName: String, entityName: String,
-                                    describeConfigEntries: List[ConfigEntry]): Unit = {
-    val optsList = List("--bootstrap-server", "localhost:9092",
-      "--entity-type", ConfigCommand.BrokerLoggerConfigType,
-      "--alter",
-      "--entity-name", entityName,
-      "--add-config", "kafka.log.LogCleaner=DEBUG",
-      "--delete-config", "kafka.server.ReplicaManager,kafka.server.KafkaApi")
-    val alterOpts = new ConfigCommandOptions(optsList.toArray)
-    var alteredConfigs = false
-
-    val resource = new ConfigResource(ConfigResource.Type.BROKER_LOGGER, resourceName)
-    val future = new KafkaFutureImpl[util.Map[ConfigResource, Config]]
-    future.complete(util.Collections.singletonMap(resource, new Config(describeConfigEntries.asJava)))
-    val describeResult: DescribeConfigsResult = mock(classOf[DescribeConfigsResult])
-    when(describeResult.all()).thenReturn(future)
-
-    val alterFuture = new KafkaFutureImpl[Void]
-    alterFuture.complete(null)
-    val alterResult: AlterConfigsResult = mock(classOf[AlterConfigsResult])
-    when(alterResult.all()).thenReturn(alterFuture)
-
-    val mockAdminClient = new MockAdminClient(util.Collections.singletonList(node), node) {
-      override def describeConfigs(resources: util.Collection[ConfigResource], options: DescribeConfigsOptions): DescribeConfigsResult = {
-        assertEquals(1, resources.size)
-        val resource = resources.iterator.next
-        assertEquals(ConfigResource.Type.BROKER_LOGGER, resource.`type`)
-        assertEquals(resourceName, resource.name)
-        describeResult
-      }
-
-      override def incrementalAlterConfigs(configs: util.Map[ConfigResource, util.Collection[AlterConfigOp]], options: AlterConfigsOptions): AlterConfigsResult = {
-        assertEquals(1, configs.size)
-        val entry = configs.entrySet.iterator.next
-        val resource = entry.getKey
-        val alterConfigOps = entry.getValue
-        assertEquals(ConfigResource.Type.BROKER_LOGGER, resource.`type`)
-        assertEquals(3, alterConfigOps.size)
-
-        val expectedConfigOps = List(
-          new AlterConfigOp(new ConfigEntry("kafka.log.LogCleaner", "DEBUG"), AlterConfigOp.OpType.SET),
-          new AlterConfigOp(new ConfigEntry("kafka.server.ReplicaManager", ""), AlterConfigOp.OpType.DELETE),
-          new AlterConfigOp(new ConfigEntry("kafka.server.KafkaApi", ""), AlterConfigOp.OpType.DELETE)
-        )
-        assertEquals(expectedConfigOps, alterConfigOps.asScala.toList)
-        alteredConfigs = true
-        alterResult
-      }
-    }
-    ConfigCommand.alterConfig(mockAdminClient, alterOpts)
-    assertTrue(alteredConfigs)
-    verify(describeResult).all()
-  }
-
-  @Test
-  def shouldSupportCommaSeparatedValuesUsingZookeeper(): Unit = {
-    val createOpts = new ConfigCommandOptions(Array("--zookeeper", zkConnect,
-      "--entity-name", "my-topic",
-      "--entity-type", "topics",
-      "--alter",
-      "--add-config", "a=b,c=[d,e ,f],g=[h,i]"))
-
-    val zkClient = mock(classOf[KafkaZkClient])
-    when(zkClient.getEntityConfigs(anyString(), anyString())).thenReturn(new Properties())
-
-    class TestAdminZkClient(zkClient: KafkaZkClient) extends AdminZkClient(zkClient) {
-      override def changeTopicConfig(topic: String, configChange: Properties): Unit = {
-        assertEquals("my-topic", topic)
-        assertEquals("b", configChange.get("a"))
-        assertEquals("d,e ,f", configChange.get("c"))
-        assertEquals("h,i", configChange.get("g"))
-      }
-    }
-
-    ConfigCommand.alterConfigWithZk(null, createOpts, new TestAdminZkClient(zkClient))
-  }
 
   @Test
   def shouldNotUpdateBrokerConfigIfMalformedEntityNameUsingZookeeper(): Unit = {
